# [COST-5427] Utilize an initial cte to optimize the distribution sql. 