--- a/koku/masu/database/sql/openshift/cost_model/distribute_cost/distribute_platform_cost.sql
+++ b/koku/masu/database/sql/openshift/cost_model/distribute_cost/distribute_platform_cost.sql
@@ -7,10 +7,41 @@ WHERE lids.usage_start >= {{start_date}}::date
 
 {% if populate %}
 WITH cte_narrow_dataset as (
-    SELECT * FROM {{schema | sqlsafe}}.reporting_ocpusagelineitem_daily_summary AS lids
+    SELECT
+        lids.usage_start,
+        lids.source_uuid,
+        lids.cost_category_id,
+        lids.pod_effective_usage_cpu_core_hours,
+        lids.pod_effective_usage_memory_gigabyte_hours,
+        lids.infrastructure_raw_cost,
+        lids.infrastructure_markup_cost,
+        lids.cost_model_cpu_cost,
+        lids.cost_model_memory_cost,
+        lids.cost_model_volume_cost,
+        lids.report_period_id,
+        lids.cluster_id,
+        lids.cluster_alias,
+        lids.data_source,
+        lids.usage_end,
+        lids.namespace,
+        lids.node,
+        lids.resource_id,
+        lids.node_capacity_cpu_cores,
+        lids.node_capacity_cpu_core_hours,
+        lids.node_capacity_memory_gigabytes,
+        lids.node_capacity_memory_gigabyte_hours,
+        lids.cluster_capacity_cpu_core_hours,
+        lids.cluster_capacity_memory_gigabyte_hours,
+        cat.name as category_name
+    FROM {{schema | sqlsafe}}.reporting_ocpusagelineitem_daily_summary AS lids
+    LEFT JOIN {{schema | sqlsafe}}.reporting_ocp_cost_category AS cat
+        ON lids.cost_category_id = cat.id
     WHERE lids.usage_start >= {{start_date}}::date
         AND lids.usage_start <= {{end_date}}::date
         AND lids.report_period_id = {{report_period_id}}
+        AND lids.namespace != 'Worker unallocated'
+        AND lids.namespace != 'Storage unattributed'
+        AND lids.namespace != 'Network unattributed'
 ),
 platform_cost AS (
     SELECT SUM(
@@ -24,9 +55,7 @@ platform_cost AS (
         filtered.source_uuid,
         filtered.cluster_id
     FROM cte_narrow_dataset as filtered
-    JOIN {{schema | sqlsafe}}.reporting_ocp_cost_category AS cat
-        ON filtered.cost_category_id = cat.id
-    WHERE cat.name = 'Platform'
+    WHERE category_name = 'Platform'
     GROUP BY filtered.usage_start, filtered.cluster_id, filtered.source_uuid
 ),
 user_defined_project_sum as (
@@ -36,10 +65,7 @@ user_defined_project_sum as (
         usage_start,
         source_uuid
     FROM cte_narrow_dataset as filtered
-    LEFT JOIN {{schema | sqlsafe}}.reporting_ocp_cost_category AS cat
-        ON filtered.cost_category_id = cat.id
-    WHERE filtered.namespace not in ('Worker unallocated', 'Platform unallocated', 'Storage unattributed', 'Network unattributed')
-        AND (cost_category_id IS NULL OR cat.name != 'Platform')
+    WHERE (filtered.cost_category_id IS NULL OR filtered.category_name != 'Platform')
     GROUP BY usage_start, cluster_id, source_uuid
 ),
 cte_line_items as (
@@ -59,19 +85,19 @@ cte_line_items as (
         max(node_capacity_memory_gigabyte_hours) as node_capacity_memory_gigabyte_hours,
         max(cluster_capacity_cpu_core_hours) as cluster_capacity_cpu_core_hours,
         max(cluster_capacity_memory_gigabyte_hours) as cluster_capacity_memory_gigabyte_hours,
-        CASE WHEN {{distribution}} = 'cpu' AND (cost_category_id IS NULL OR max(cat.name) != 'Platform') THEN
+        CASE WHEN {{distribution}} = 'cpu' AND (cost_category_id IS NULL OR max(filtered.category_name) != 'Platform') THEN
             CASE WHEN max(udps.usage_cpu_sum) <= 0 THEN
                 0
             ELSE
                 (sum(pod_effective_usage_cpu_core_hours) / max(udps.usage_cpu_sum)) * max(pc.platform_cost)::decimal
             END
-        WHEN {{distribution}} = 'memory'AND (cost_category_id IS NULL OR max(cat.name) != 'Platform') THEN
+        WHEN {{distribution}} = 'memory'AND (cost_category_id IS NULL OR max(filtered.category_name) != 'Platform') THEN
             CASE WHEN max(udps.usage_memory_sum) <= 0 THEN
                 0
             ELSE
                 (sum(pod_effective_usage_memory_gigabyte_hours) / max(udps.usage_memory_sum)) * max(pc.platform_cost)::decimal
             END
-        WHEN max(cat.name) = 'Platform' THEN
+        WHEN max(filtered.category_name) = 'Platform' THEN
             0 - SUM(
                     COALESCE(infrastructure_raw_cost, 0) +
                     COALESCE(infrastructure_markup_cost, 0) +
@@ -88,10 +114,7 @@ cte_line_items as (
     JOIN user_defined_project_sum as udps
         ON udps.usage_start = filtered.usage_start
         AND udps.cluster_id = filtered.cluster_id
-    LEFT JOIN {{schema | sqlsafe}}.reporting_ocp_cost_category AS cat
-        ON filtered.cost_category_id = cat.id
     WHERE filtered.namespace IS NOT NULL
-        AND filtered.namespace not in ('Worker unallocated', 'Storage unattributed', 'Network unattributed')
     GROUP BY filtered.usage_start, filtered.node, filtered.namespace, filtered.cluster_id, cost_category_id, filtered.data_source
 )
 INSERT INTO {{schema | sqlsafe}}.reporting_ocpusagelineitem_daily_summary (