--- a/koku/masu/database/sql/openshift/cost_model/distribute_cost/distribute_unattributed_network_cost.sql
+++ b/koku/masu/database/sql/openshift/cost_model/distribute_cost/distribute_unattributed_network_cost.sql
@@ -6,7 +6,13 @@ WHERE lids.usage_start >= {{start_date}}::date
 ;
 
 {% if populate %}
-WITH unattributed_network_cost AS (
+WITH cte_narrow_dataset as (
+    SELECT * FROM {{schema | sqlsafe}}.reporting_ocpusagelineitem_daily_summary AS lids
+    WHERE lids.usage_start >= {{start_date}}::date
+        AND lids.usage_start <= {{end_date}}::date
+        AND lids.report_period_id = {{report_period_id}}
+),
+unattributed_network_cost AS (
     SELECT SUM(
             COALESCE(infrastructure_raw_cost, 0) +
             COALESCE(infrastructure_markup_cost, 0)+
@@ -14,13 +20,10 @@ WITH unattributed_network_cost AS (
             COALESCE(cost_model_memory_cost, 0) +
             COALESCE(cost_model_volume_cost, 0)
         ) as unattributed_cost,
-        lids.usage_start
-    FROM {{schema | sqlsafe}}.reporting_ocpusagelineitem_daily_summary as lids
-    WHERE lids.usage_start >= {{start_date}}::date
-        AND lids.usage_start <= {{end_date}}::date
-        AND lids.report_period_id = {{report_period_id}}
-        AND lids.namespace = 'Network unattributed'
-    GROUP BY lids.usage_start
+        cnd.usage_start
+    FROM {{schema | sqlsafe}}.cte_narrow_dataset as cnd
+    WHERE cnd.namespace = 'Network unattributed'
+    GROUP BY cnd.usage_start
 ),
 user_defined_project_sum as (
     SELECT sum(pod_effective_usage_cpu_core_hours) as usage_cpu_sum,
@@ -28,26 +31,23 @@ user_defined_project_sum as (
         cluster_id,
         usage_start,
         source_uuid
-    FROM {{schema | sqlsafe}}.reporting_ocpusagelineitem_daily_summary as lids
+    FROM {{schema | sqlsafe}}.cte_narrow_dataset as cnd
     LEFT JOIN {{schema | sqlsafe}}.reporting_ocp_cost_category AS cat
-        ON lids.cost_category_id = cat.id
-    WHERE lids.usage_start >= {{start_date}}::date
-        AND lids.usage_start <= {{end_date}}::date
-        AND report_period_id = {{report_period_id}}
-        AND lids.namespace not in ('Worker unallocated', 'Platform unallocated', 'Storage unattributed', 'Network unattributed')
+        ON cnd.cost_category_id = cat.id
+    WHERE cnd.namespace not in ('Worker unallocated', 'Platform unallocated', 'Storage unattributed', 'Network unattributed')
         AND (cost_category_id IS NULL OR cat.name != 'Platform')
     GROUP BY usage_start, cluster_id, source_uuid
 ),
 cte_line_items as (
     SELECT
         max(report_period_id) as report_period_id,
-        lids.cluster_id,
+        cdn.cluster_id,
         max(cluster_alias) as cluster_alias,
-        lids.data_source as data_source,
-        lids.usage_start,
+        cdn.data_source as data_source,
+        cdn.usage_start,
         max(usage_end) as usage_end,
-        lids.namespace,
-        lids.node,
+        cdn.namespace,
+        cdn.node,
         max(resource_id) as resource_id,
         max(node_capacity_cpu_cores) as node_capacity_cpu_cores,
         max(node_capacity_cpu_core_hours) as node_capacity_cpu_core_hours,
@@ -78,21 +78,18 @@ cte_line_items as (
         END AS distributed_cost,
         max(cost_category_id) as cost_category_id,
         max(raw_currency) as raw_currency
-    FROM {{schema | sqlsafe}}.reporting_ocpusagelineitem_daily_summary AS lids
+    FROM {{schema | sqlsafe}}.cte_narrow_dataset as cnd
     LEFT JOIN unattributed_network_cost as usc
-        ON usc.usage_start = lids.usage_start
+        ON usc.usage_start = cdn.usage_start
     LEFT JOIN user_defined_project_sum as udps
-        ON udps.usage_start = lids.usage_start
-        AND udps.cluster_id = lids.cluster_id
+        ON udps.usage_start = cdn.usage_start
+        AND udps.cluster_id = cdn.cluster_id
     LEFT JOIN {{schema | sqlsafe}}.reporting_ocp_cost_category AS cat
-        ON lids.cost_category_id = cat.id
-    WHERE lids.usage_start >= {{start_date}}::date
-        AND lids.usage_start <= {{end_date}}::date
-        AND report_period_id = {{report_period_id}}
-        AND lids.namespace IS NOT NULL
-        AND lids.namespace not in ('Worker unallocated', 'Platform unallocated', 'Storage unattributed')
+        ON cdn.cost_category_id = cat.id
+    WHERE cdn.namespace IS NOT NULL
+        AND cdn.namespace not in ('Worker unallocated', 'Platform unallocated', 'Storage unattributed')
         AND (cost_category_id IS NULL OR cat.name != 'Platform')
-    GROUP BY lids.usage_start, lids.node, lids.namespace, lids.cluster_id, cost_category_id, lids.data_source
+    GROUP BY cdn.usage_start, cdn.node, cdn.namespace, cdn.cluster_id, cost_category_id, cdn.data_source
 )
 INSERT INTO {{schema | sqlsafe}}.reporting_ocpusagelineitem_daily_summary (
     uuid,
@@ -105,29 +102,12 @@ INSERT INTO {{schema | sqlsafe}}.reporting_ocpusagelineitem_daily_summary (
     namespace,
     node,
     resource_id,
-    pod_labels,
-    pod_usage_cpu_core_hours,
-    pod_request_cpu_core_hours,
-    pod_effective_usage_cpu_core_hours,
-    pod_limit_cpu_core_hours,
-    pod_usage_memory_gigabyte_hours,
-    pod_request_memory_gigabyte_hours,
-    pod_effective_usage_memory_gigabyte_hours,
-    pod_limit_memory_gigabyte_hours,
     node_capacity_cpu_cores,
     node_capacity_cpu_core_hours,
     node_capacity_memory_gigabytes,
     node_capacity_memory_gigabyte_hours,
     cluster_capacity_cpu_core_hours,
     cluster_capacity_memory_gigabyte_hours,
-    persistentvolumeclaim,
-    persistentvolume,
-    storageclass,
-    volume_labels,
-    persistentvolumeclaim_capacity_gigabyte,
-    persistentvolumeclaim_capacity_gigabyte_months,
-    volume_request_storage_gigabyte_months,
-    persistentvolumeclaim_usage_gigabyte_months,
     source_uuid,
     cost_model_rate_type,
     distributed_cost,
@@ -145,29 +125,12 @@ SELECT
     ctl.namespace,
     ctl.node,
     ctl.resource_id,
-    NULL as pod_labels,
-    NULL as pod_usage_cpu_core_hours,
-    NULL as pod_request_cpu_core_hours,
-    NULL as pod_effective_usage_cpu_core_hours,
-    NULL as pod_limit_cpu_core_hours,
-    NULL as pod_usage_memory_gigabyte_hours,
-    NULL as pod_request_memory_gigabyte_hours,
-    NULL as pod_effective_usage_memory_gigabyte_hours,
-    NULL as pod_limit_memory_gigabyte_hours,
     ctl.node_capacity_cpu_cores,
     ctl.node_capacity_cpu_core_hours,
     ctl.node_capacity_memory_gigabytes,
     ctl.node_capacity_memory_gigabyte_hours,
     ctl.cluster_capacity_cpu_core_hours,
     ctl.cluster_capacity_memory_gigabyte_hours,
-    NULL as persistentvolumeclaim,
-    NULL as persistentvolume,
-    NULL as storageclass,
-    NULL as volume_labels,
-    NULL as persistentvolumeclaim_capacity_gigabyte,
-    NULL as persistentvolumeclaim_capacity_gigabyte_months,
-    NULL as volume_request_storage_gigabyte_months,
-    NULL as persistentvolumeclaim_usage_gigabyte_months,
     UUID '{{source_uuid | sqlsafe}}' as source_uuid,
     'unattributed_network' as cost_model_rate_type,
     ctl.distributed_cost,