… koku/masu/database/sql/openshift/cost_model/distribute_cost/distribute_unattributed_storage_cost.sql
… koku/masu/database/sql/openshift/cost_model/distribute_cost/distribute_worker_cost.sql