@lcouzens 2024-10-04
I tried loading some large customer data for this but I don't think I can scale large enough to really see the impact. That said based on the explains its certainly not worse! I say we roll with it and take the improvement!